modifySelectionText -> mapSelectionText
