+++ lib/get-selection-text.js
@@ -1,7 +1,7 @@
-import iterateSelectionText from './iterate-selection-text';
+import mapSelectionText from './map-selection-text';
 
 export default articleJson => {
   const result = [];
-  iterateSelectionText(articleJson, text => result.push(text));
+  mapSelectionText(articleJson, text => result.push(text));
   return result;
 };
