+++ lib/index.js
@@ -1,9 +1,9 @@
 import _getSelectionItems from './get-selection-items';
 import _getSelectionText from './get-selection-text';
 import _mapSelectionItems from './map-selection-items';
-import _modifySelectionText from './modify-selection-text';
+import _mapSelectionText from './map-selection-text';
 
 export const getSelectionItems = _getSelectionItems;
 export const getSelectionText = _getSelectionText;
 export const mapSelectionItems = _mapSelectionItems;
-export const modifySelectionText = _modifySelectionText;
+export const mapSelectionText = _mapSelectionText;
